[fli-iam#2059] Divide VIP api controller by endpoint
