+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/vip/controller/PipelineApi.java
@@ -1,0 +1,66 @@
+/**
+
+ * Shanoir NG - Import, manage and share neuroimaging data
+ * Copyright (C) 2009-2019 Inria - https://www.inria.fr/
+ * Contact us on https://project.inria.fr/shanoir/
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
+ */
+
+package org.shanoir.ng.vip.controller;
+
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.Parameter;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import org.shanoir.ng.shared.core.model.IdName;
+import org.shanoir.ng.shared.exception.EntityNotFoundException;
+import org.shanoir.ng.shared.exception.RestServiceException;
+import org.shanoir.ng.shared.exception.SecurityException;
+import org.shanoir.ng.vip.dto.VipExecutionDTO;
+import org.shanoir.ng.vip.monitoring.model.ExecutionStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+
+import java.io.IOException;
+
+/**
+ * @author Alae Es-saki
+ */
+@Tag(name = "VIP pipeline", description="Proxy API for VIP /pipelines")
+@RequestMapping("/vip/pipeline")
+public interface PipelineApi {
+    @Operation(summary = "Get all available pipelines in VIP", description = "Returns all the pipelines available to the authenticated user in VIP", tags={  })
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "successful response, returns the status"),
+            @ApiResponse(responseCode = "403", description = "forbidden"),
+            @ApiResponse(responseCode = "500", description = "unexpected error"),
+            @ApiResponse(responseCode = "503", description = "error from VIP API")})
+    @RequestMapping(value = "",
+            produces = { "application/json", "application/octet-stream" },
+            method = RequestMethod.GET)
+    ResponseEntity<String> getPipelineAll() throws SecurityException;
+
+    @Operation(summary = "Get the description of pipeline [name] in the version [version]", description = "Returns the VIp description of the pipeline [name] in the version [version].", tags={  })
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "successful response, returns the status"),
+            @ApiResponse(responseCode = "403", description = "forbidden"),
+            @ApiResponse(responseCode = "500", description = "unexpected error"),
+            @ApiResponse(responseCode = "503", description = "error from VIP API")})
+    @RequestMapping(value = "/{identifier}/{version}",
+            produces = { "application/json", "application/octet-stream" },
+            method = RequestMethod.GET)
+    ResponseEntity<String> getPipeline(@Parameter(name = "The pipeline identifier", required=true) @PathVariable("identifier") String identifier, @Parameter(name = "The pipeline version", required=true) @PathVariable("version") String version) throws SecurityException;
+
+
+}
